Removed the AzureBaseServices.cs class
