+++ AIAssistView/AIAssistView.csproj
@@ -63,7 +63,6 @@
 		<PackageReference Include="Microsoft.Maui.Controls" Version="$(MauiVersion)" />
 		<PackageReference Include="Microsoft.Maui.Controls.Compatibility" Version="$(MauiVersion)" />
 		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="*" />
-		<PackageReference Include="Microsoft.SemanticKernel" Version="*" />
 		<PackageReference Include="Syncfusion.Maui.AIAssistView" Version="*" />
 	</ItemGroup>
